fitnessApp/requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
